data_sources/requirements.txt
geopandas
pyarrow
google-cloud-bigquery